+++ NEWS.md
@@ -20,6 +20,8 @@
 
 * Fixed a bug in `getGvE` with multiple traits.
 
+* Clarified in function documentation that `simParam = NULL` uses the global `SP` object where applicable.
+
 # AlphaSimR 2.1.0
 
 * changed R6 and methods from Depends to Imports to match current best practices for R packages